Doc-Scanner/requirements.txt
surya-ocr
opencv-python
numpy
Pillow
python-docx
python-docx-replace